pyproject.toml
[project]
name = "animalharmbench-testbed"
version = "0.1.0"
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "datasets>=4.4.1",
    "google-genai>=1.56.0",
    "inspect-ai>=0.3.150",
    "inspect-evals", # This fork pins AHB to version 2.1
    "numpy>=2.2.6",
    "openai>=2.14.0",
    "osfclient>=0.0.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "tqdm>=4.67.1",
    "trl[peft]>=0.25.1",
    "vllm>=0.11.0",
    "wandb>=0.23.1",
]

[project.optional-dependencies]
flashinfer = [ # Beware: Doesn't work on Blackwell GPUs yet: https://github.com/flashinfer-ai/flashinfer/issues/1353
    "flashinfer-cubin>=0.6.0",
    "flashinfer-python>=0.6.0",
]
flash-attn = [
    "flash-attn",
]

[dependency-groups]
dev = [
    "ruff>=0.14.4",
]

[tool.uv]
constraint-dependencies = [
    "torch==2.8.*", # Because flash_attn depends on it.
]

[tool.uv.sources]
flash-attn = { path = "flash_attn-2.8.3+cu128torch2.8-cp313-cp313-linux_x86_64.whl" }
torch = [
  { index = "pytorch-cu128" },
]
inspect-evals = { git = "https://github.com/lukasgebhard/inspect_evals", branch = "ahb-2-0" } # Use "ahb-2-1" for AHB-2.1

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
